Refactor home sections into components
